+++ Core/OfficeDevPnP.Core/Sites/SiteCollection.cs
@@ -1,5 +1,4 @@
-﻿
-#if !SP2013 && !SP2016
+﻿#if !SP2013 && !SP2016
 #if !ONPREMISES
 using Microsoft.Graph;
 #endif
@@ -7,6 +6,7 @@ using Microsoft.SharePoint.Client;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using OfficeDevPnP.Core.Diagnostics;
+using OfficeDevPnP.Core.Entities;
 using OfficeDevPnP.Core.Utilities;
 using OfficeDevPnP.Core.Utilities.Async;
 using System;
@@ -20,7 +20,6 @@ using System.Text.Encodings.Web;
 #endif
 using System.Threading.Tasks;
 using System.Web;
-using System.Linq;
 
 namespace OfficeDevPnP.Core.Sites
 {
@@ -303,7 +302,7 @@ namespace OfficeDevPnP.Core.Sites
             int maxRetryCount = 12, // Maximum number of retries (12 x 10 sec = 120 sec = 2 mins)
             int retryDelay = 1000 * 10 // Wait time default to 10sec,
 #if !SP2019
-            , 
+            ,
             bool noWait = false
 #endif
             )
@@ -558,12 +557,58 @@ namespace OfficeDevPnP.Core.Sites
 
             if (group != null && !string.IsNullOrEmpty(group.SiteUrl))
             {
+                // Try to configure the site/group classification, if any
+                if (!string.IsNullOrEmpty(siteCollectionCreationInformation.Classification))
+                {
+                    await SetTeamSiteClassification(
+                        siteCollectionCreationInformation.Classification,
+                        group.GroupId,
+                        graphAccessToken
+                        );
+                }
+
                 responseContext = clientContext.Clone(group.SiteUrl);
             }
 
             return responseContext;
         }
 #endif
+
+        private static async Task SetTeamSiteClassification(string classification, string groupId, string graphAccessToken)
+        {
+            // Patch the created group
+            using (var handler = new HttpClientHandler())
+            {
+                using (var httpClient = new PnPHttpProvider(handler))
+                {
+                    string requestUrl = $"https://graph.microsoft.com/v1.0/groups/{groupId}";
+
+                    // Serialize request object to JSON
+                    var jsonBody = JsonConvert.SerializeObject(new { classification });
+                    var requestBody = new StringContent(jsonBody);
+
+                    // Build Http request
+                    HttpRequestMessage request = new HttpRequestMessage(new HttpMethod("PATCH"), requestUrl);
+                    request.Content = requestBody;
+                    if (MediaTypeHeaderValue.TryParse("application/json", out MediaTypeHeaderValue jsonMediaType))
+                    {
+                        requestBody.Headers.ContentType = jsonMediaType;
+                    }
+                    if (!string.IsNullOrEmpty(graphAccessToken))
+                    {
+                        request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", graphAccessToken);
+                    }
+
+                    // Perform actual post operation
+                    HttpResponseMessage response = await httpClient.SendAsync(request, new System.Threading.CancellationToken());
+
+                    if (!response.IsSuccessStatusCode)
+                    {
+                        throw new Exception("Failed to set Classification for created group");
+                    }
+                }
+            }
+        }
 
 #if !ONPREMISES
         /// <summary>
@@ -1364,13 +1409,13 @@ namespace OfficeDevPnP.Core.Sites
         /// Enable Microsoft Teams team in an O365 group connected team site
         /// Will also enable it on a newly Groupified classic site
         /// </summary>
-        /// <param name="context"></param>
+        /// <param name="context">Context to operate against</param>
         /// <returns></returns>
         public static async Task<string> TeamifySiteAsync(ClientContext context)
         {
             string responseString = null;
 
-            context.Site.EnsureProperties(s => s.GroupId);
+            context.Site.EnsureProperty(s => s.GroupId);
 
             if (context.Web.IsSubSite())
             {
@@ -1486,7 +1531,7 @@ namespace OfficeDevPnP.Core.Sites
 
             await context.Web.ExecutePost("/_api/sitepages/communicationsite/enable", $@" {{ ""designPackageId"": ""{designPackageId.ToString()}"" }}");
         }
-#endif
+
 
         /// <summary>
         /// Get sensitivity label id for a given Label
@@ -1573,6 +1618,45 @@ namespace OfficeDevPnP.Core.Sites
             }
         }
 
+        /// <summary>
+        /// Deletes a Communication site or a Group-less Modern team site.
+        /// </summary>
+        /// <param name="context">Context to operate against</param>
+        /// <returns></returns>
+        public static async Task<bool> DeleteSiteAsync(ClientContext context)
+        {
+            bool siteDeleted = false;
+
+            var webTemplateId = context.Web.GetBaseTemplateId();
+
+            context.Site.EnsureProperties(s => s.Id, s => s.GroupId, s => s.Url);
+
+            if (webTemplateId == "SITEPAGEPUBLISHING#0" || webTemplateId == "STS#3")
+            {
+                var result = await context.Web.ExecutePost("/_api/SPSiteManager/delete", $@" {{ ""siteId"": ""{context.Site.Id.ToString()}"" }}");
+
+                var parsedResult = JObject.Parse(result);
+
+                siteDeleted = Convert.ToBoolean(parsedResult["odata.null"]);
+
+                return await Task.Run(() => siteDeleted);
+            }
+            else if (webTemplateId == "GROUP#0" || context.Site.GroupId != Guid.Empty)
+            {
+                var result = await context.Web.ExecutePost($"/_api/GroupSiteManager/Delete?siteUrl='{context.Site.Url}'", string.Empty);
+
+                var parsedResult = JObject.Parse(result);
+
+                siteDeleted = Convert.ToBoolean(parsedResult["odata.null"]);
+
+                return await Task.Run(() => siteDeleted);
+            }
+            else
+            {
+                throw new Exception("Only deletion of Communication site or Modern team site is supported by this method.");
+            }
+        }
+#endif
     }
 }
 #endif